Move CheckPantherParser to testutil
