+++ internal/log_analysis/log_processor/parsers/awslogs/cloudtrail_insights_test.go
@@ -25,7 +25,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/parsers"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/parsers/testutil"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/parsers/timestamp"
 )
@@ -155,26 +154,12 @@ func TestCloudTrailInsightParser(t *testing.T) {
 	expectedEventEnd.AppendAnyAWSAccountIds("123456789012")
 	expectedEventEnd.SetEvent(expectedEventEnd)
 	expectedEventStart.SetEvent(expectedEventStart)
-	CheckPantherParser(t, log, &CloudTrailInsightParser{}, expectedEventStart.Log(), expectedEventEnd.Log())
+	testutil.CheckPantherParser(t, log, &CloudTrailInsightParser{}, expectedEventStart.Log(), expectedEventEnd.Log())
 }
 
 func TestCloudTrailInsightLogType(t *testing.T) {
 	parser := &CloudTrailInsightParser{}
 	require.Equal(t, "AWS.CloudTrailInsight", parser.LogType())
-}
-
-func CheckPantherParser(t *testing.T, log string, parser parsers.LogParser, expect *parsers.PantherLog, expectMore ...*parsers.PantherLog) {
-	t.Helper()
-	p := parser.New()
-	results := p.Parse(log)
-	require.NotNil(t, results)
-	// Prepend the required log arg to more
-	expectMore = append([]*parsers.PantherLog{expect}, expectMore...)
-	require.Equal(t, len(expectMore), len(results), "Invalid number of pather logs produced by parser")
-	for i, result := range results {
-		expect := expectMore[i]
-		testutil.EqualPantherLog(t, expect, []*parsers.PantherLog{result})
-	}
 }
 
 // func CheckJSONEqual(t *testing.T, expect interface{}, actual interface{}, msgAndArgs ...interface{}) {
